requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
pydantic
psycopg2-binary
python-dotenv